# api/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for the KFamily family-tree app. Every /api/v1 endpoint except /auth/signin needs a Casdoor access token.",
        "title": "KFamily API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/audit-history/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Every change recorded for a record (any table's primary key), newest first. Each entry holds the full row right after the change.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit history"
                ],
                "summary": "List audit history for a record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source record id (primary key)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/services_audithistory.Dto"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.HttpResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.HttpResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.HttpResult"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signin": {
            "post": {
                "description": "Rate limited per client IP.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in with a Casdoor authorization code",
                "parameters": [
                    {
                        "description": "Code and state from Casdoor's redirect",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services_auth.SigninInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.HttpResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/services_auth.TokenDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.HttpResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.HttpResult"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/util.HttpResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "services_audithistory.Dto": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "create | update | delete",
                    "type": "string"
                },
                "changedBy": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "id": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        },
        "services_auth.SigninInputDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "services_auth.TokenDto": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "expiresIn": {
                    "type": "integer"
                },
                "refreshToken": {
                    "type": "string"
                },
                "tokenType": {
                    "type": "string"
                }
            }
        },
...
        "util.HttpResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "isSuccess": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "result": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Bearer \u003cCasdoor access token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}